Set the correct path to the mtrouter.
