--- framework/mcrouter.vcxproj.filters
@@ -1,126 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="Source Files">
-      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
-      <Extensions>cpp;c;cc;cxx;def;odl;idl;hpj;asm;asmx</Extensions>
-    </Filter>
-    <Filter Include="Header Files">
-      <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
-      <Extensions>h;hpp;hxx;hm;inl;inc;xsd</Extensions>
-    </Filter>
-    <Filter Include="Resource Files">
-      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
-      <Extensions>bat;rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="system\GEPlatform.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mtrouter\resources\resource.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mtrouter\resources\targetver.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mtrouter\app\MultitargetRouter.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mtrouter\app\NEMultitargetRouterSettings.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mtrouter\service\RouterServerService.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mtrouter\service\private\ServiceRegistry.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mtrouter\service\private\ListServiceProxies.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mtrouter\service\private\ServiceProxy.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mtrouter\service\private\ServiceStub.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mtrouter\app\private\RouterConsoleService.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="mtrouter\resources\mtrouter.ico">
-      <Filter>Resource Files</Filter>
-    </None>
-    <None Include="mtrouter\resources\mtrouter.service.install.bat">
-      <Filter>Resource Files</Filter>
-    </None>
-    <None Include="mtrouter\resources\mtrouter.service.uninstall.bat">
-      <Filter>Resource Files</Filter>
-    </None>
-    <None Include="mtrouter\resources\small.ico">
-      <Filter>Resource Files</Filter>
-    </None>
-    <None Include="mtrouter\resources\Readme.md">
-      <Filter>Resource Files</Filter>
-    </None>
-    <None Include="mtrouter\resources\mtrouter.service">
-      <Filter>Resource Files</Filter>
-    </None>
-    <None Include="mtrouter\resources\mtrouter_post_build.bat">
-      <Filter>Resource Files</Filter>
-    </None>
-    <None Include="mtrouter\Readme.md">
-      <Filter>Resource Files</Filter>
-    </None>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="mtrouter\app\private\MultitargetRouter.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="mtrouter\service\private\RouterServerService.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="mtrouter\service\private\ServiceRegistry.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="mtrouter\service\private\ListServiceProxies.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="mtrouter\service\private\ServiceProxy.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="mtrouter\service\private\ServiceStub.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="mtrouter\app\private\NEMultitargetRouterSettings.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="mtrouter\app\private\posix\MultitargetRouterPosix.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="mtrouter\app\private\win32\MultitargetRouterWin32.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="mtrouter\app\private\RouterConsoleService.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <ResourceCompile Include="mtrouter\resources\mtrouter.rc">
-      <Filter>Resource Files</Filter>
-    </ResourceCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <Text Include="mtrouter\app\private\CMakeLists.txt">
-      <Filter>Resource Files</Filter>
-    </Text>
-    <Text Include="mtrouter\service\private\CMakeLists.txt">
-      <Filter>Resource Files</Filter>
-    </Text>
-    <Text Include="mtrouter\CMakeLists.txt">
-      <Filter>Resource Files</Filter>
-    </Text>
-  </ItemGroup>
-</Project>
