+++ framework/mtrouter/resources/mtrouter.service.uninstall.bat
